--- about.txt
@@ -1,21 +1,0 @@
-id : A notation for a house
-date: Date house was sold
-price: Price is prediction target
-bedrooms: Number of bedrooms
-bathrooms: Number of bathrooms
-sqft_living: Square footage of the home
-sqft_lot: Square footage of the lot
-floors :Total floors (levels) in house
-waterfront :House which has a view to a waterfront
-view: Has been viewed
-condition :How good the condition is overall
-grade: overall grade given to the housing unit, based on King County grading system
-sqft_above : Square footage of house apart from basement
-sqft_basement: Square footage of the basement
-yr_built : Built Year
-yr_renovated : Year when house was renovated
-zipcode: Zip code
-lat: Latitude coordinate
-long: Longitude coordinate
-sqft_living15 : Living room area in 2015(implies-- some renovations) This might or might not have affected the lotsize area
-sqft_lot15 : LotSize area in 2015(implies-- some renovations)
