+++ statistical_calculations/__init__.py
@@ -1,0 +1,3 @@
+import csv
+import numpy as np
+import matplotlib.pyplot as plt
