--- views.py
@@ -1,4 +1,0 @@
-from main import reading_csv
-
-data_array = reading_csv()
-print(data_array)
